// apps/storefront/prismicio-types.d.ts
export interface LookbookSliceDefaultItem {
    /**
     * Context field in *Lookbook → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Woman
     * - **API ID Path**: lookbook.items[].context
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    context: prismic.SelectField<'Woman' | 'Man', 'filled'>;

    /**
     * Desktop Image field in *Lookbook → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Lookbook → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Heading field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].heading
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    heading: prismic.KeyTextField;

    /**
     * Product Handle 1 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_handle_1
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_handle_1: prismic.KeyTextField;

    /**
     * Product Title 1 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_title_1
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_title_1: prismic.KeyTextField;

    /**
     * Variant Title 1 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].variant_title_1
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    variant_title_1: prismic.KeyTextField;

    /**
     * Product Handle 2 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_handle_2
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_handle_2: prismic.KeyTextField;

    /**
     * Product Title 2 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_title_2
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_title_2: prismic.KeyTextField;

    /**
     * Variant Title 2 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].variant_title_2
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    variant_title_2: prismic.KeyTextField;

    /**
     * Product Handle 3 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_handle_3
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_handle_3: prismic.KeyTextField;

    /**
     * Product Title 3 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_title_3
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_title_3: prismic.KeyTextField;

    /**
     * Variant Title 3 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].variant_title_3
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    variant_title_3: prismic.KeyTextField;

    /**
     * Product Handle 4 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_handle_4
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_handle_4: prismic.KeyTextField;

    /**
     * Product Title 4 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].product_title_4
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_title_4: prismic.KeyTextField;

    /**
     * Variant Title 4 field in *Lookbook → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: lookbook.items[].variant_title_4
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    variant_title_4: prismic.KeyTextField;
}

/**
 * Default variation for Lookbook Slice
 *
 * - **API ID**: `default`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type LookbookSliceDefault = prismic.SharedSliceVariation<
    'default',
    Simplify<LookbookSliceDefaultPrimary>,
    Simplify<LookbookSliceDefaultItem>
>;

/**
 * Slice variation for *Lookbook*
 */
type LookbookSliceVariation = LookbookSliceDefault;

/**
 * Lookbook Shared Slice
 *
 * - **API ID**: `lookbook`
 * - **Description**: Lookbook
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type LookbookSlice = prismic.SharedSlice<
    'lookbook',
    LookbookSliceVariation
>;